requirements.txt
pillow==10.2.0
pyserial==3.5
websockets==12.0
